refactor player module
